requirements.txt
ccxt
requests
pandas
numpy
matplotlib
transformers
yfinance
stable-baselines3
gym
schedule
python-dotenv
